Fix #24958: Fixed inline forms using UUID-PK parents with auto-PK children
The fix for issue #24377 fixed using inline forms with parent and
child objects that use UUIDs as primary keys, but did not cover
the case where a parent object used a UUID for a primary key along
with a child object that used an AutoField for a primary key.
It also fixes the case where a parent object has an AutoField
primary key with a UUID alternate key, and the child object is
related to the parent's alternate key.